+++ ntex-bytes/src/string.rs
@@ -182,6 +182,18 @@ impl<T: AsRef<str>> PartialEq<T> for ByteString {
     }
 }
 
+impl PartialEq<ByteString> for &ByteString {
+    fn eq(&self, other: &ByteString) -> bool {
+        self.as_str() == other.as_str()
+    }
+}
+
+impl PartialEq<ByteString> for &str {
+    fn eq(&self, other: &ByteString) -> bool {
+        *self == other.as_str()
+    }
+}
+
 impl AsRef<str> for ByteString {
     #[inline]
     fn as_ref(&self) -> &str {