+++ ntex-util/src/error.rs
@@ -40,6 +40,11 @@ impl ErrorMessageChained {
         }
     }
 
+    /// Construct ErrorMessage from ErrorMessageChained
+    pub const fn from_bstr(msg: ByteString) -> Self {
+        Self { msg, source: None }
+    }
+
     pub fn msg(&self) -> &ByteString {
         &self.msg
     }
@@ -50,8 +55,14 @@ impl ErrorMessageChained {
 }
 
 impl ErrorMessage {
+    /// Construct new empty ErrorMessage
     pub const fn empty() -> Self {
         Self(ByteString::from_static(""))
+    }
+
+    /// Construct ErrorMessage from ByteString
+    pub const fn from_bstr(msg: ByteString) -> ErrorMessage {
+        ErrorMessage(msg)
     }
 
     pub fn is_empty(&self) -> bool {
@@ -117,5 +128,45 @@ impl<M: Into<ErrorMessage>> From<M> for ErrorMessageChained {
             msg: value.into().0,
             source: None,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn error_message() {
+        let msg = ErrorMessage::empty();
+        assert!(msg.is_empty());
+        assert_eq!(msg.as_str(), "");
+        assert_eq!(msg.as_bstr(), ByteString::new());
+        assert_eq!(ByteString::new(), msg.as_bstr());
+        assert_eq!(ByteString::new(), msg.into_string());
+
+        let msg = ErrorMessage::from("test");
+        assert!(!msg.is_empty());
+        assert_eq!(msg.as_str(), "test");
+        assert_eq!(msg.as_bstr(), ByteString::from("test"));
+
+        let msg = ErrorMessage::from("test".to_string());
+        assert!(!msg.is_empty());
+        assert_eq!(msg.as_str(), "test");
+        assert_eq!(msg.as_bstr(), ByteString::from("test"));
+
+        let msg = ErrorMessage::from(ByteString::from("test"));
+        assert!(!msg.is_empty());
+        assert_eq!(msg.as_str(), "test");
+        assert_eq!(msg.as_bstr(), ByteString::from("test"));
+
+        assert_eq!(ByteString::from(&msg), "test");
+        assert_eq!(ByteString::from(msg), "test");
+    }
+
+    #[test]
+    fn error_message_chained() {
+        let chained = ErrorMessageChained::from(ByteString::from("test"));
+        assert_eq!(chained.msg(), "test");
+        assert!(chained.source().is_none());
     }
 }